Fix the CI build failures of the branch
The CI builds treat warnings as errors (WARNINGS_AS_ERRORS=yes), which the local builds did not:

- TcpEndpointI::acceptor: the adapter name is only used by the Network.framework listener; mark the parameter
  unused for the other platforms, where GCC and MSVC fail on it.
- TcpEndpointI.cpp: only include the BSD-socket acceptor and transceiver headers where they are used. On
  Network.framework the Acceptor and Transceiver interfaces declare the asynchronous operations, which these final
  classes do not implement there, and clang flags an abstract final class.
- ThreadPool: _promote belongs to the readiness-based implementation; move it into that block, it is unused with
  IOCP and Network.framework.
- WSTransceiver::startRead: make the pointer-difference to size conversions explicit, the function is now compiled
  with clang.
- NetworkFrameworkTLS: SecTrustGetCertificateAtIndex is deprecated since macOS 12 and iOS 15. Take the leaf
  certificate from SecTrustCopyCertificateChain, available on the deployment targets of the branch.
- Format the files clang-format flagged.

+++ cpp/src/Ice/WSTransceiver.cpp
@@ -770,11 +770,11 @@ IceInternal::WSTransceiver::startRead(Buffer& buf)
             // no more than the payload length. The remaining of the buffer will be
             // sent over in another frame.
             //
-            size_t readSz = _readPayloadLength - (buf.i - _readStart);
+            size_t readSz = _readPayloadLength - static_cast<size_t>(buf.i - _readStart);
             if (static_cast<size_t>(buf.b.end() - buf.i) > readSz)
             {
                 size_t size = buf.b.size();
-                buf.b.resize(buf.i - buf.b.begin() + readSz);
+                buf.b.resize(static_cast<size_t>(buf.i - buf.b.begin()) + readSz);
                 _delegate->startRead(buf);
                 buf.b.resize(size);
             }
